[Refactor] 공사일보 화면 구조 분리 및 로직 정리
